[🚚rename]: useKit 파일 경로 위치 이동
